backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.1

# Vector Database
pgvector==0.2.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# AI & Embeddings
openai>=1.10.0
langchain>=0.1.5
langchain-openai>=0.0.5
# tiktoken installed as dependency of langchain-openai
faiss-cpu==1.9.0.post1

# Data Processing
pandas==2.2.0
numpy==1.26.3
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2

# HTTP & Scraping
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# Validation
pydantic==2.10.4
pydantic-settings==2.1.0

# Utilities
python-json-logger==2.0.7
tenacity==8.2.3
